My-Skills/code-logic/scripts/requirements.txt
pyyaml>=6.0
tree-sitter>=0.22.0
tree-sitter-rust>=0.21.0
tree-sitter-python>=0.21.0
networkx>=3.0
graphviz>=0.20
argparse
